requirements.txt
aws-cdk.core
aws-cdk.assertions
aws-cdk.aws_ec2
aws-cdk.aws_eks